Speed up unloading trie time
Originally to unload the trie the program simply iterated over every
pointer of the entire trie and free them bottom up, and this worked
perfectly fine when the trie only had 26 pointers; however, after
adding support for unicode the trie had tens of thousands of pointers
making it untenable to continue doing this without wasting a ton of
time. So to address this, the program simply keeps an unordered list
of every trie node created then unloads that at the end, and this means
the unload time only is affected by the number of nodes this makes it
magnitudes faster when using the full unicode set of characters.