Track highlighting through booleans instead of case
Previously all characters were stored and checked in lower
case, and then set to upper case if they needed to be highlighted.
Now characters are stored in bchars or boolean characters, where
the boolean determines if it needs to be highlighted allowing for
future support of case sensitive searchs and allow characters
that dont have upper or lower cases to be checked for.